Moved presentation out of the index
